U7 data type

+++ src/types.rs
@@ -1,4 +1,4 @@
 pub(crate) type Frequency = f64;
 pub(crate) type Cents = f64;
 pub(crate) type Semitones = f64;
-pub(crate) type Octave = u8;
+pub(crate) type Octave = i32;
